requirements.txt
asgiref==3.7.2
Django==5.0.2
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
asgiref==3.7.2
django-classy-tags==4.1.0
django-sekizai==4.1.0
django-widget-tweaks==1.5.0
sqlparse==0.4.4
psycopg2-binary==2.9.5
djangorestframework==3.14.0

